# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go + Gin baseline for a learning-focused streaming system.",
        "title": "demo-streaming API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/stream-keys": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-keys"
                ],
                "summary": "Create stream key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.StreamKeyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stream-keys/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-keys"
                ],
                "summary": "Refresh stream key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.StreamKeyResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/stream-keys/revoke": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stream-keys"
                ],
                "summary": "Revoke stream key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.RevokeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/streamkeys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "streamkeys.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "streamkeys.IngestInfo": {
            "type": "object",
            "properties": {
                "provider": {
                    "type": "string"
                },
                "rtmp_url": {
                    "type": "string"
                }
            }
        },
        "streamkeys.RevokeResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "streamkeys.StreamKeyResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "ingest": {
                    "$ref": "#/definitions/streamkeys.IngestInfo"
                },
                "stream_key": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}